parse/pyproject.toml
[project]
name = "parse"
version = "0.1.0"
description = ""
authors = [
    {name = "Park Kyeong Jun",email = "100195725+Kyeong6@users.noreply.github.com"}
]
readme = "README.md"
packages = [ ]
package-mode = false
requires-python = "^3.11"
dependencies = [
    "langchain>=0.3.20",
    "langchain-core>=0.3.43",
    "python-dotenv>=1.0.1",
    "langchain-openai>=0.3.8",
    "langgraph>=0.3.5",
    "fastapi>=0.115.11",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
    "langsmith>=0.3.13",
    "ipython (>=9.2.0,<10.0.0)",
    "anthropic (>=0.50.0,<0.51.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "lxml (>=5.4.0,<6.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "langchain-upstage (>=0.6.0,<0.7.0)",
    "pinecone (==6.0.2)",
    "pinecone-plugin-interface (==0.0.7)",
    "pinecone-text (==0.10.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "pytz (>=2025.2,<2026.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"